WID-225. Refactor and add some more tests for data wrapper; move tests related to TS widget into their own folder
